fix broken context menu "open file location"
After an user reported the issue in discord, I noticed some weird behavior too when using a portable version and trying to use "open file location" on an source image, where it kept sending me to my user documents folder and not the actual file location. Target video location seems to be fine, but I just applied the same fix for both because it's a robust solution and now opens the file locations reliably for both.